[Tab Group Sync] Migration of local IDs from Java to native
This CL migrates tab group ID mappings stored in Java shared prefs to
native ModelTypeStore. The shared prefs is read on startup and cleared
out while the data is passed to the sync bridge. The sync bridge
subsequently during database load completes the migration by adding
the ID mapping to the stored entries and writing it back.

Note, we don't do any code cleanup here for TabGroupStore. Instead
we replace it with EmptyTabGroupStoreDelegate which effectively
disables that code path.
We also start observing ID update events in the sync bridge which
then writes the mapping to ModelTypeStore.

The migration will run on subsequent startup. The shared prefs is
cleared out, hence the migration runs only once. The migration is
behind a disabled feature flag. Hence there is no functional change
with this CL. It will be enabled in a subsequent CL.

TabGroupMetadataPersistentStore.java is soon to be deleted and
covered via Java native unittests, but the tool doesn't recognize
the coverage.

Low-Coverage-Reason: COVERAGE_UNDERREPORTED The reported file
Bug: b:343268785
Change-Id: Ic1736d8003d2a2e6e46d9973a3183a2462a2f452
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5583203
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Commit-Queue: Shakti Sahu <shaktisahu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1309143}