[views-ax] Use ViewAccessibility::SetRole in //chrome
This CL migrates callers using View::SetAccessibleRole to instead use
the new ViewAccessibility::SetRole setter for all files in //chrome.
This change does not introduce any behavior change and is just a rename
done as part of the ViewsAX project. Here is the design doc for the
ViewsAX project:
https://docs.google.com/document/d/1Ku7HOyDsiZem1yaV6ccZ-tz3lO2XR2NEcm8HjR6d-VY/edit#heading=h.ke1u3utej413

This CL does not migrate the SetAccessibleRole method for
views::Builder. A follow-up CL will take care of that part.

Bug: 325137417Change-Id: I49a4d8a025a6ffe51455924de8d2cce4be5e5955
Change-Id: I49a4d8a025a6ffe51455924de8d2cce4be5e5955
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5588156
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Commit-Queue: Ashish Kumar <ashishkum@microsoft.com>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1309180}