Remove DumpWithoutCrashing in EmbeddedWorkerInstanceClientImpl
This CL removes DwoC code from
EmbeddedWorkerInstanceClientImpl::StartWorker(). This was introduced in
crrev.com/c/5568543, but we have to remove it before going to Dev
because it produces too many crashes.

We already got lots of crash reports. That should be enough for
debugging. Also we have UMA to record the length of
cors_exempt_header_list. That will help future debugging as well.

Bug: 40753993, 343752126
Change-Id: I8d6a7c94e77603f065cece7eb5162f3c6bd61d42
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5588023
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1309183}