[text-box-trim] Fix two block-in-inline in a row
When the first formatted line is inside of a block-in-inline,
`IsBlockStartTrimmed` needs to be propagated from the block.

When normal inline follows, the propagation isn't necessary
because the algorithm checks `IsFirstFormattedLine()`, but the
issue appears when a block-in-inline follows it. The test for
followed by normal inline is in:
`text-box-trim-block-in-inline-start-001`.

Tests for end will be in following patches.

All changes are under a runtime flag enabled only for testing.

Bug: 40254880
Change-Id: I5f053d5999d98378ef4b39cacec2d6923aa1d8c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5591191
Reviewed-by: Lingqi Chi <lingqi@chromium.org>
Auto-Submit: Koji Ishii <kojii@chromium.org>
Commit-Queue: Lingqi Chi <lingqi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1309185}