[views-ax] Use ViewAccessibility::SetRole in //ui
This CL migrates callers using View::SetAccessibleRole to instead use
the new ViewAccessibility::SetRole setter for all files in //ui. This
change does not introduce any behavior change and is just a rename
done as part of the ViewsAX project. Here is the design doc for the
ViewsAX project:
https://docs.google.com/document/d/1Ku7HOyDsiZem1yaV6ccZ-tz3lO2XR2NEcm8HjR6d-VY/edit#heading=h.ke1u3utej413

This CL does not migrate the SetAccessibleRole method for
views::Builder. A follow-up CL will take care of that part.

Bug: 325137417
Change-Id: Ifcfeecccf692ad33ad0b3fc5f0922d16cd670606
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5588157
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Commit-Queue: Ashish Kumar <ashishkum@microsoft.com>
Reviewed-by: Dirk Pranke <dpranke@google.com>
Cr-Commit-Position: refs/heads/main@{#1309203}