[//gpu] Have WebGPUDecoderImpl pass internal usage to BeginScopedAccess
This CL has WebGPUDecoderImpl pass the client-provided internal usage
to the Dawn representation's BeginScopedAccess(), as well as passing it
in to the point where it will be used in the Skia fallback when we get
to the point of actually switching over to using client-provided
internal usages. Note that BeginScopedAccess() does not yet actually
*do* anything with `internal_usage`; this is still part of wiring
things up end-to-end.

Remaining links in the chain to change Dawn representations from setting
hardcoded internal usages to using client-provided internal usages:

* Modify callers of AssociateMailbox() to pass appropriate internal
  usages
* Pass internal usages down to individual Dawn reps via BeginAccess()
* Under a killswitch, change individual Dawn reps to use
  client-supplied internal usages instead of their current
  hardcoded internal usages

Bug: 339171225
Change-Id: I2035851e91d5c3ee2e0ec59c8f6a17002da111ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5589107
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1309215}