Fix flakiness in testMigrateSyncToSignin custom passphrase EG tests
Before this CL: These tests turned on sync, then added an encrypted
item to the FakeServer to get the client into the "passphrase required"
state. However, they didn't properly wait for the client to notice the
change (they triggered an update, but didn't wait for it to finish).

After this CL: The encrypted item is added to the server before even
turning on sync. This way, the client cannot miss it.

Bug: 342972808
Change-Id: I05d4a854d07a06d279461751baf064afda95ddaf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5587868
Commit-Queue: Marc Treib <treib@chromium.org>
Reviewed-by: Jood Hajeer <jood@google.com>
Cr-Commit-Position: refs/heads/main@{#1309233}