[sync] Use base::debug::Alias() to debug slowness for sessions on iOS
Some of the modified codepaths are involved in the crash (slowness)
reports in the linked bug. The code looks good, so the current theory
is that those reports must be running into a large number of tabs,
tracked sessions or both.

Change-Id: Iec763f542e58063353673a4dd9f06af51e3e6f3f
Bug: 41483250
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5587838
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1309236}