[mullet M2] Move 'Fill everything' at the top of the context sub-menu
This CL adds a variation to the granular filling flag, which allows
moving the 'Fill everything' button at the top of the context menu
(instead of it being in the footer). Also, when at the top,
'Fill everything' will always be displayed (including in full form
filling mode).

This CL is part of an effort to split the mullet M2 project (the
granular filling project) in as many flag variations as reasonably
possible.
Mullet M2 has shown metrics regressions, and by separating it into
multiple experiments, we are trying to identify which part of the
feature is at fault.

There seem to be no consequences to moving this button out of the
footer. `IsFooterSuggestionType()` returning false for this button will
not influence any behavior, it looks like it only changes the style of
the button (which is desired, we want this button to have the same style
as 'Fill full name', not the same style as footer buttons).

Demo: https://screenshot.googleplex.com/6A9BFWxZXA6GG6h
      https://screenshot.googleplex.com/4YQW8Ac4U6SVaR6

Bug: 40274514
Change-Id: I9fc1aa22007e7f792b16ba08b828f744fdbb30c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5576195
Reviewed-by: Bruno Braga <brunobraga@google.com>
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Commit-Queue: Theo Cristea <theocristea@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1309243}