[TrustedVault] Fix bad request errors (400) when joining security domain
This patch fixes a bug by reverting to the behavior before
https://crrev.com/c/5329654 accidentally started populating the request
proto with `MEMBER_TYPE_PHYSICAL_DEVICE` instead of the desired
`MEMBER_TYPE_UNSPECIFIED`, when the caller uses
`UnspecifiedAuthenticationFactorType` (formerly known as
`AuthenticationFactorType::kUnspecified`).

Without this fix, the request runs into a 400 error returned by the
server, which in practice means users that are in degraded
recoverability state cannot resolve the error UI, even if they follow
the required steps and reauth entering their account password.

To avoid similar issues in the future, the fake server used in
integration tests (`FakeSecurityDomainsServer`) is updated to return an
error to the client in similar circumnstances.

Change-Id: I828ff3634b8d9dc9e60178fd9d4d8f28c5a59638
Fixed: 344347963
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5591730
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Maksim Moskvitin <mmoskvitin@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1309248}