[Reauth] Impl AndroidxDeviceAuthenticatorControllerImpl#authenticate
Androidx.biometric.BiometricPrompt needs a fragment activity so that it
could attach the UI of the biometric auth to this activity. For this
reason, the activity needs to be passed into
AndroidxDeviceAuthenticatorControllerImpl. All the usages of the
biometric authentication feature were updated appropriately.

increase. This is the library we prefer to use for biometric
authentication and unfortunately the size increase can't be avoided.
redirects calls to the androidx.biometric library and has no its own
logic to test.

Bug: 340473460
Change-Id: I676da51d615eae7ad710f409cc6a0651ae9fdb9e
Binary-Size: androidx.biometric library usage has caused the binary size
Low-Coverage-Reason: AndroidxDeviceAuthenticatorControllerImpl only
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5563248
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Commit-Queue: Anna Tsvirchkova <atsvirchkova@google.com>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1309273}