[//gpu] Have Mac Dawn rep use client's internal usages based on Feature
We are changing Dawn's SharedImage representations to use internal
usages specified by the client when creating textures rather than
hardcoding these. This CL introduces a base::Feature that we will use as
a killswitch when making this change and changes the Mac representation.
As the changes to some of the representations are non-trivial based on
what their texture creation paths look like, I did not attempt to change
all of the ~10 relevant texture creation points to use this Feature. I
will do so in followup CLs.

Bug: 339171225
Change-Id: Ie3d51d52733d3b2684bce610fe7713c3213acd60
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5588083
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1309284}