[iOS] Make passkeys visible in the password manager
The SavedPasswordsPresenter already supports receiving a
webauthn::PasskeyModel object to present passkeys. This CL
provides the PasskeyModel to the SavedPasswordsPresenter
on iOS, which allows visualizing the passkeys in the
Password Manager.

Also added a feature check to make sure the PasskeyModel
object is only ever be non nil if the kSyncWebauthnCredentials
feature is enabled.

Low-Coverage-Reason: OTHER Code is still hidden behind the kSyncWebauthnCredentials feature flag and other pieces of the feature are still missing.
Bug: 330355124
Change-Id: I02e148059ebf48aad94cbf9c44e0785838716698
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5582373
Commit-Queue: Alexis Hétu <sugoi@chromium.org>
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1309366}