part1: FlexLayoutView respect SizeBounds
When FlexLayout respects the available space, the size constraints are
passed to the TabbedPaneTabStrip. The default flex of the horizontal
TabbedPaneTabStrip is 1. This means that it always takes up all the
space in the horizontal direction. However, from the perspective of the
translated bubble, this does not seem to be the case. Let's ignore its
horizontal space constraints for now.

Bug: 40232718
Change-Id: Icb1f872a0f9fa02a0bce6c976085c1acbdfe6726
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5587543
Commit-Queue: Weidong Liu <weidongliu@chromium.org>
Reviewed-by: Allen Bauer <kylixrd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1309370}