Uprev ChromeScreenAI CIPD to 124.3
This CL updates the CIPD for downloading the library that is used for
ChromeSreenAI browser tests.
If browser tests fail, rervet the rollout for the new binary in the
component updater as well. You will need to rollout a newer (or older)
release to update clients that are affected by the faulty rollout.

Change-Id: I36d8cc96c080e2ccfa774f56a0ce3a14480e1ad9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5589546
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: Kyungjun Lee <kyungjunlee@google.com>
Commit-Queue: Ramin Halavati <rhalavati@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1309407}