Add the contents of //build/cros_cache/common/ to Simple Chrome tests
Simple Chrome will download a gsutil installation at common/ in the
//build/cros_cache/. gsutil shouldn't be used at all during test-time
for VM tests. But for some HW tests, gsutil is used to download the
DUT image on-the-fly. Without this gsutil in the test's inputs,
cros-flash will fetch and manage its own gsutil installation.

In crbug.com/341991534, we're seeing some weird behavior in that
installation. But just re-using the gsutil we already have deployed
in the cros_cache appears to resolve all that. So this just puts
all of common/ in the test inputs for simple chrome tests. In local
checkouts, common/ is only 20-30 MB. Compared to the overall cache
size of 7-8 GB, this shouldn't make much of a difference.

This will then need https://crrev.com/c/5587435 for cros-flash to
pick up the cache.

Bug: 341991534
Change-Id: Ia8a6aa62566479ddf68ba1f91d87c03e57f6e079
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5590456
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Sven Zheng <svenzheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1309417}