Set `threat_source` when local hash checks pass synchronously
When a URL is allowlisted, real time mechanisms fall back to the local
hash based checks. If those checks pass synchronously, we return a SAFE
threat type with no threat source. But that's not quite accurate because
we did do a local hash based check. Similarly, when the local hash based
check is the default and passes synchronously, we have no threat
source. This affects SafeBrowsing.CheckUrl.FirstRequestThreatType.

This CL changes the threat source to improve metrics. Because the threat
type is always SAFE in this situation, it should have no other impact.

OBSOLETE_HISTOGRAMS=Moved to SafeBrowsing.CheckUrl.FirstRequestThreatType2

Bug: b/330359965
Change-Id: I85b8a286fe9b16b0d97f7c96745d58bd4b467f70
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5586102
Reviewed-by: thefrog <thefrog@chromium.org>
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1309451}