ozone: Use atomic calls for setting connector props.
Rename SetProperty to SetConnectorProperty to make clear that it's
targeting connector props.

Setting properties was done using the legacy uapi which we don't want to
use often.
Similar to setting CRTC properties, use an atomic commit to set
connector properties.
This doesn't introduce any functionality changes as the legacy uapi uses
atomic under the hood for atomic drivers.
The atomic commit is blocking to match the same behavior of the legacy
call.

This change is to pave the way of serializing atomic calls.

BUG=b:342617770
TEST=unittests using this call + chrome logs with props being set

Change-Id: Iaffd83a6d9f959929b4a1d15b24ca579e298761a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5577274
Reviewed-by: Gil Dekel <gildekel@chromium.org>
Commit-Queue: Mark Yacoub <markyacoub@google.com>
Cr-Commit-Position: refs/heads/main@{#1309469}