Specify ownership for CreateTestWidget() callers in chrome/browser
We are changing Views so that the default (and only) ownership model
for Widgets will be CLIENT_OWNS_WIDGET. However, there are APIs like
CreateTestWidget() that rely on a default value for ownership. We
need to specify the ownership explicitly so that we can then track
down and change all of the uses as part of the refactoring.

This CL updates all the callers in //chrome/browser that are not in
//chrome/browser/ui/views (which are updated in a separate CL).

There should be no functional changes in this CL.

Bug: 339619005
Change-Id: I5485c35b8b177169328dfb218f6ae071f24f4ff0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5586637
Reviewed-by: Prameet Shah <phshah@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1309481}