[Tab Group Sync] Correct implementation for sign-out behavior
This CL implements proper behavior for sign-out. The new
behavior is to close out all the groups and the associated tabs except
the ones that were created before sign-in. This is in conformance with
the new guidelines for sign-in and sync. The earlier behavior was to
close all groups on sign out.

As part of this, we are storing a boolean
created_before_syncing_tab_groups as a local-only field in the
ModelTypeStore. This field is set on SavedTabGroup creation and never
mutated after that.

Bug: 342618065
Change-Id: Ic94d3ad683b76a5396b75b2d974e01c56dcfb8ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5590124
Commit-Queue: Shakti Sahu <shaktisahu@chromium.org>
Reviewed-by: Darryl James <dljames@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1309544}