[Read aloud] Fix enum collision and add tests.
This fixes an enum logic collision that I intended to address when I
wrote the code but ultimately forgot to. Enums are converted to
numbers during runtime, and VoicePackServerStatusSuccessCode.NOT_INSTALLED will evaluate to equal
VoicePackServerStatusErrorCode.OTHER, and so on for the rest of the
values. This is why I wrapped the enums in
VoicePackServerResponseSuccess and VoicePackServerResponseError
objects.

This CL breaks up the switch statement into two, one of for
each interface.

Bug: 344038789
Change-Id: I757af3caaf9d5f362d70e7530f1abacaf4e2f423
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5593020
Commit-Queue: Eitan Goldberger <eitang@google.com>
Reviewed-by: Kristi Saney <kristislee@google.com>
Cr-Commit-Position: refs/heads/main@{#1309579}