Simplifications to LensSearchBubbleController
This CL is a refactor with no intended behavior change.

Previously, LensSearchBubbleController inherited from BrowserUserData.
This caused communication between LensSearchBubbleController and
LensOverlayController to indirect through Browser. Given that the former
was 1 per browser window and the latter is 1 per tab, this was fragile.

This CL causes LensOverlayController to explicitly own
LensSearchBubbleController. This CL deletes
LensSearchBubbleInteractiveUiTest as the test does not make sense --
there is no way to invoke a standalone LensSearchBubble. The existing
logic is tested via CloseSearchBubbleOnOverlayInteraction.

Change-Id: I3281e8229bb64682b95e11a26aa7e34febcd4562
Bug: 339497734
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5590643
Commit-Queue: Erik Chen <erikchen@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1309588}