Handle IllegalArgumentException in FlatBuffer deserialization
We're seeing a small number of IllegalArgumentExceptions in
FlatBuffer deserialization. I suspect these are due to an I/O
error. Note, not 100% of legacy TabState restores are successful -
there are fallback paths such as reloading the URL from the
Tab metadata file. In this case, we fail to restore the
WebContentsState, the TabState isn't usable (and we should
fallback to reloading from the Tab metadata file). This is
because TabImpl assumes a frozen Tab has a non-null WebContentsState
(see line 1843 in TabImpl - see go/frozen-webcontentsstate-null).

Bug: b:344629383
Change-Id: I4ae7d1790cbd91c2c64d5a81622997d5d2eab58b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5594769
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: David Maunder <davidjm@chromium.org>
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1309595}