[User Education] Fix flaky lifecycle test.
There was a potential race condition between checking for window
activation and observing window activation; this was one frame but in
ubsan/msan builds it was slow enough to cause the activation to slip in
between.

I need to do a follow-up to better address the underlying cause of this
issue, which is that the widget focus observer cannot currently detect
initial state (i.e. the initially-focused widget).

Bug: 342251760
Change-Id: I660df82eb986373a2dd6c08e19648eb647430271
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5593927
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1309612}